feat: improve basic auth method
The extract_user_credentials used to assume that its
`decoded_base64_authorization_header` argument only had one ':' to
separate the user email and password. Now it allows the user password to
contain any number of ':' characters.
It is still assumed that an email will have no such characters in it
(:).