feat: user stored in request
In app.py, the user instance of an authorized request is stored in the
request as `request.current_user`. This allows easier access to the user
instance.
In users.py, the route `/users/<user_id>`, now allows the user's
to_json() data to be returned using the user instance from
request.current_user.