add: multiple functions for auth
Private function `_generate_uuid` returns a random uuid which will be used for generating a session id upon authenticated login.
`valid_login` checks the authenticity of the given credentials.
`create_session` is responsible for creating and assigning a session_id for the user upon successful login. This id is attached to the user and saved to the database.