add: flask app with various feats
Simple flask app is defined. The `AUTH` constant is used to access the auth.py file's public authentication methods. The `/index` route returns a simple json payload. The `/users` route creates a new user.
The `/sessions` route authenticates a user's login credentials - If successful, the user is logged in and assigned a session_id. The session_id attached to the user is saved to the database.
If unsuccessful, control is given to the 401 status abort.